eatfood/pyproject.toml
[tool.poetry]
name = "eatfood"
version = "0.1.0"
description = ""
authors = ["Neeraj319 <itskop520@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-crispy-forms = "^1.14.0"
Pillow = "^9.0.0"
django-ckeditor = "^6.2.0"

[tool.poetry.dev-dependencies]
yes = "^0.0.1"
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
